<Fix:> suggestion 에서 update 오류 수정 <Feat:> update profile 에서 닉네임 중복 문제 해결

+++ project/backend/domain/suggestion/suggestion_router.py
@@ -14,7 +14,9 @@ router = APIRouter(
 
 
 @router.delete("/remove/{suggest_id}")
-async def remove_suggest(suggest_id: int, db: Session = Depends(get_db)):
+async def remove_suggest(suggest_id: int,
+                         current_user: User = Depends(get_current_user),
+                         db: Session = Depends(get_db)):
     """
      등록한 의견 삭제 : 27page 2번
       - 입력예시 :id = 1
@@ -22,30 +24,34 @@ async def remove_suggest(suggest_id: int, db: Session = Depends(get_db)):
     suggest = suggestion_crud.get_suggest(db, id=suggest_id)
     if suggest is None:
         raise HTTPException(status_code=404, detail="suggestion not found")
+    if suggest.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="you can't remove this suggestion")
     db.delete(suggest)
     db.commit()
     return {"detail": "Suggest deleted successfully"}
 
 
-@router.patch("/update/{suggest_id}")
+@router.patch("/update/{suggest_id}", response_model=suggestion_schema.SuggestionSchema)
 async def update_suggest(suggest_id: int, title: str = Form(...), content: str = Form(...),
+                         current_user: User = Depends(get_current_user),
                          db: Session = Depends(get_db)):
     """
      등록한 의견 수정 : 27page 2번
       - 입력예시 :id = 1, title: 집가고싶어, content: 집가고싶어
      """
-    suggest = suggestion_crud.get_suggest(db, suggest_id=suggest_id)
+    suggest = suggestion_crud.get_suggest(db, suggest_id)
     if suggest is None:
         raise HTTPException(status_code=404, detail="suggestion not found")
+    if current_user.id != suggest.user_id:
+        raise HTTPException(status_code=403, detail="you are not the owner of this suggestion")
     suggest.title = title
     suggest.content = content
-    db.add(suggest)
     db.commit()
     db.refresh(suggest)
-    return {"suggest": suggest}
+    return suggest
 
 
-@router.get("/get/{suggest_id}/text", response_model=suggestion_schema.Suggestion_content_schema)
+@router.get("/get/{suggest_id}/text", response_model=suggestion_schema.SuggestionContentSchema)
 def get_Suggest_id(suggest_id: int, db: Session = Depends(get_db)):
     """
     개발자 의견제출  : 27page 1번
@@ -58,7 +64,7 @@ def get_Suggest_id(suggest_id: int, db: Session = Depends(get_db)):
     return suggest
 
 
-@router.post("/post", status_code=status.HTTP_204_NO_CONTENT)
+@router.post("/post", response_model=suggestion_schema.SuggestionSchema)
 def post_Suggest(current_user: User = Depends(get_current_user), title: str = Form(...), content: str = Form(...),
                  db: Session = Depends(get_db)):
     """
@@ -74,10 +80,10 @@ def post_Suggest(current_user: User = Depends(get_current_user), title: str = Fo
     db.add(new_suggest)
     db.commit()
     db.refresh(new_suggest)
-    return {"suggest": new_suggest}
+    return new_suggest
 
 
-@router.get("/get/all_title", response_model=List[suggestion_schema.Suggestion_title_schema])
+@router.get("/get/all_title", response_model=List[suggestion_schema.SuggestionTitleSchema])
 def get_Suggest_all(current_user: User = Depends(get_current_user), db: Session = Depends(get_db)):
     """
     개발자 의견제출  : 27page 3번
